requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
python-dotenv
streamlit